+++ libs/proto/job.go
@@ -3,7 +3,7 @@ package proto
 // TODO optimize struct after replace kafka
 type KafkaMsg struct {
 	OP       string   `json:"op"`
-	RoomId   int32    `json:"roomid,omitempty"`
+	RoomId   int64    `json:"roomid,omitempty"`
 	ServerId int32    `json:"server,omitempty"`
 	SubKeys  []string `json:"subkeys,omitempty"`
 	Msg      []byte   `json:"msg"`
